Pin the physics: invariant and transform tests, and an honest T
Adds the test tier the review found missing: rule 110's exact first
rows, the blinker's period-2 cycle, Gray-Scott bounded in [0, 1], Lenia
running in 3D end to end, and a vmap-over-shared-module rollout. T is
annotated float — it is a time resolution, and the Lenia example itself
passes 1/3.

+++ src/cax/cs/flow_lenia/update.py
@@ -32,7 +32,7 @@ class FlowLeniaUpdate(LeniaUpdate):
 		self,
 		channel_size: int,
 		*,
-		T: int,
+		T: float,
 		growth_fn: Callable = exponential_growth_fn,
 		rule_params: LeniaRuleParams,
 		# Flow Lenia parameters
