+++ src/cax/cs/lenia/cs.py
@@ -30,7 +30,7 @@ class Lenia(ComplexSystem[Array, Array]):
 		channel_size: int,
 		*,
 		R: int,
-		T: int,
+		T: float,
 		state_scale: float = 1.0,
 		kernel_fn: Callable = gaussian_kernel_fn,
 		growth_fn: Callable = exponential_growth_fn,
