+++ src/cax/cs/lenia/update.py
@@ -31,7 +31,7 @@ class LeniaUpdate(Update[Array, Array, Array]):
 		self,
 		channel_size: int,
 		*,
-		T: int,
+		T: float,
 		growth_fn: Callable = exponential_growth_fn,
 		rule_params: LeniaRuleParams,
 	):
